Started refactoring core to make easier to understand fns.
The idea is that we can query the game-state to ask certain
things about the state of our game. Previously this was
always done by passing the board directly but this is often
incorrect and additional state needs to calculated and passed
in to ask the query you want. This leaves the users of the API
with too much detailed knowledge of the game.

Given any state of the game, wether during or at the start of
a new turn we should be able to query the game state to give
us a list of possible moves, validate a certain move or turn we
want to play at that point etc. We should not have to be concerned
with whose turn it actually is or whether we are playing the first
turn of a game or another.